@@ -5,6 +5,7 @@ declare(strict_types=1);
 namespace CodeLieutenant\LaravelCrypto;
 
 use CodeLieutenant\LaravelCrypto\Console\GenerateCryptoKeysCommand;
+use CodeLieutenant\LaravelCrypto\Contracts\Encoder;
 use CodeLieutenant\LaravelCrypto\Contracts\Hashing;
 use CodeLieutenant\LaravelCrypto\Contracts\KeyLoader;
 use CodeLieutenant\LaravelCrypto\Contracts\PublicKeySigning;
@@ -33,16 +34,19 @@ use CodeLieutenant\LaravelCrypto\Signing\Hmac\Blake2b as HmacBlake2b;
 use CodeLieutenant\LaravelCrypto\Signing\Hmac\Sha256 as HmacSha256;
 use CodeLieutenant\LaravelCrypto\Signing\Hmac\Sha512 as HmacSha512;
 use CodeLieutenant\LaravelCrypto\Signing\SigningManager;
+use CodeLieutenant\LaravelCrypto\Support\Random;
 use Illuminate\Contracts\Config\Repository;
 use Illuminate\Contracts\Encryption\Encrypter;
 use Illuminate\Contracts\Foundation\Application;
 use Illuminate\Encryption\Encrypter as LaravelConcreteEncrypter;
 use Illuminate\Encryption\EncryptionServiceProvider;
+use Override;
 use Psr\Log\LoggerInterface;
+use Random\Engine\Secure;
+use Random\Randomizer;
 
 class ServiceProvider extends EncryptionServiceProvider
 {
-
     public function boot(): void
     {
         if ($this->app->runningInConsole()) {
@@ -50,6 +54,7 @@ class ServiceProvider extends EncryptionServiceProvider
         }
     }
 
+    #[Override]
     public function register(): void
     {
         if ($this->app->runningInConsole()) {
@@ -76,14 +81,15 @@ class ServiceProvider extends EncryptionServiceProvider
         ];
 
         foreach ($encoders as $encoder) {
-            $this->app->singleton($encoder, function (Application $app) use ($encoder) {
-                $config = $app->make(Repository::class)->get('crypto.encoder.config.' . $encoder);
+            $this->app->singleton($encoder, function (Application $app) use ($encoder): PhpEncoder|JsonEncoder|MessagePackEncoder|IgbinaryEncoder {
+                $config = $app->make(Repository::class)->get('crypto.encoder.config.'.$encoder);
+
                 return new $encoder(...$config);
             });
         }
 
         $this->app->singleton(
-            Contracts\Encoder::class,
+            Encoder::class,
             $this->app->make(Repository::class)->get('crypto.encoder.driver')
         );
     }
@@ -92,21 +98,21 @@ class ServiceProvider extends EncryptionServiceProvider
     {
         $this->app->singleton(
             AppKeyLoader::class,
-            fn(Application $app) => AppKeyLoader::make($app->make(Repository::class))
+            fn (Application $app): AppKeyLoader => AppKeyLoader::make($app->make(Repository::class))
         );
         $this->app->singleton(
             Blake2BHashingKeyLoader::class,
-            fn(Application $app) => Blake2BHashingKeyLoader::make($app->make(Repository::class))
+            fn (Application $app): Blake2BHashingKeyLoader => Blake2BHashingKeyLoader::make($app->make(Repository::class))
         );
 
         $this->app->singleton(
             HmacKeyLoader::class,
-            fn(Application $app) => HmacKeyLoader::make($app->make(Repository::class))
+            fn (Application $app): HmacKeyLoader => HmacKeyLoader::make($app->make(Repository::class))
         );
 
         $this->app->singleton(
             EdDSASignerKeyLoader::class,
-            fn(Application $app) => EdDSASignerKeyLoader::make(
+            fn (Application $app): EdDSASignerKeyLoader => EdDSASignerKeyLoader::make(
                 $app->make(Repository::class),
                 $app->make(LoggerInterface::class)
             )
@@ -128,17 +134,15 @@ class ServiceProvider extends EncryptionServiceProvider
         ];
 
         foreach ($hmacSigners as $signer) {
-            $this->app->singleton($signer, function (Application $app) use ($signer) {
-                $config = $app->make(Repository::class)->get('crypto.signing.config.' . $signer);
+            $this->app->singleton($signer, function (Application $app) use ($signer): \CodeLieutenant\LaravelCrypto\Signing\Hmac\Blake2b|\CodeLieutenant\LaravelCrypto\Signing\Hmac\Sha256|\CodeLieutenant\LaravelCrypto\Signing\Hmac\Sha512 {
+                $config = $app->make(Repository::class)->get('crypto.signing.config.'.$signer);
                 $keyLoader = $app->make(HmacKeyLoader::class);
 
                 return $config !== null ? new $signer($keyLoader, $config) : new $signer($keyLoader);
             });
         }
 
-        $this->app->singleton(Signing::class, static function (Application $app) {
-            return $app->make($app->make(Repository::class)->get('crypto.signing.driver'));
-        });
+        $this->app->singleton(Signing::class, static fn (Application $app) => $app->make($app->make(Repository::class)->get('crypto.signing.driver')));
 
         $this->app->singleton(PublicKeySigning::class, EdDSA::class);
     }
@@ -152,25 +156,24 @@ class ServiceProvider extends EncryptionServiceProvider
         ];
 
         foreach ($hashers as $hasher) {
-            $this->app->singleton($hasher, static function (Application $app) use ($hasher) {
-                $params = $app->make(Repository::class)->get('crypto.hashing.config.' . $hasher);
+            $this->app->singleton($hasher, static function (Application $app) use ($hasher): Blake2b|Sha256|Sha512 {
+                $params = $app->make(Repository::class)->get('crypto.hashing.config.'.$hasher);
 
-                return $params === null ? new $hasher() : new $hasher(...$params);
+                return $params === null ? new $hasher : new $hasher(...$params);
             });
         }
 
-        $this->app->singleton(Hashing::class, static function (Application $app) {
-            return $app->make($app->make(Repository::class)->get('crypto.hashing.driver'));
-        });
+        $this->app->singleton(Hashing::class, static fn (Application $app) => $app->make($app->make(Repository::class)->get('crypto.hashing.driver')));
 
         $this->app->singleton(HashingManager::class);
     }
 
     protected function getConfigPath(): string
     {
-        return __DIR__ . '/../config/crypto.php';
+        return __DIR__.'/../config/crypto.php';
     }
 
+    #[Override]
     protected function registerEncrypter(): void
     {
         foreach ([AesGcm256Encrypter::class, XChaCha20Poly1305Encrypter::class] as $encryptor) {
@@ -181,7 +184,7 @@ class ServiceProvider extends EncryptionServiceProvider
         }
 
         $func = static function (Application $app) {
-            $cipher = $app->make('config')->get('app.cipher');
+            $cipher = $app->make(Repository::class)->get('app.cipher');
 
             $enc = Encryption::tryFrom($cipher);
 
@@ -192,9 +195,15 @@ class ServiceProvider extends EncryptionServiceProvider
             return match ($enc) {
                 Encryption::SodiumAES256GCM => $app->make(AesGcm256Encrypter::class),
                 Encryption::SodiumXChaCha20Poly1305 => $app->make(XChaCha20Poly1305Encrypter::class),
+                Encryption::AES256GCM, Encryption::AES256CBC => new LaravelConcreteEncrypter(
+                    key: $app->make(AppKeyLoader::class)->getKey(),
+                    cipher: $enc->value,
+                ),
             };
         };
 
+        $this->app->singleton(Randomizer::class, fn (): Randomizer => new Randomizer(new Secure));
+        $this->app->singleton(Random::class, fn (): Random => new Random($this->app->make(Randomizer::class)));
         $this->app->singleton(Encrypter::class, $func);
         $this->app->singleton('encrypter', $func);
     }
