@@ -4,6 +4,4 @@ declare(strict_types=1);
 
 namespace CodeLieutenant\LaravelCrypto\Contracts;
 
-interface PublicKeySigning extends Signing
-{
-}
+interface PublicKeySigning extends Signing {}
