@@ -5,30 +5,27 @@ declare(strict_types=1);
 namespace CodeLieutenant\LaravelCrypto\Keys\Generators;
 
 use CodeLieutenant\LaravelCrypto\Contracts\KeyGenerator;
-use CodeLieutenant\LaravelCrypto\Encryption\AesGcm256Encrypter;
-use CodeLieutenant\LaravelCrypto\Encryption\XChaCha20Poly1305Encrypter;
 use CodeLieutenant\LaravelCrypto\Enums\Encryption;
 use CodeLieutenant\LaravelCrypto\Traits\EnvKeySaver;
 use Illuminate\Contracts\Config\Repository;
 use Illuminate\Encryption\Encrypter;
 
-class AppKeyGenerator implements KeyGenerator
+final readonly class AppKeyGenerator implements KeyGenerator
 {
     use EnvKeySaver;
 
-    protected const CONFIG_CIPHER_PATH = 'app.cipher';
-    protected const CONFIG_KEY_PATH = 'app.key';
+    private const string CONFIG_CIPHER_PATH = 'app.cipher';
 
-    public const ENV = 'APP_KEY';
+    private const string CONFIG_KEY_PATH = 'app.key';
 
-    public function __construct(protected readonly Repository $config)
-    {
-    }
+    public const string ENV = 'APP_KEY';
+
+    public function __construct(private Repository $config) {}
 
     public function generate(?string $write): ?string
     {
-        $old = $this->config->get(static::CONFIG_KEY_PATH);
-        $cipher = $this->config->get(static::CONFIG_CIPHER_PATH);
+        $old = $this->config->get(self::CONFIG_KEY_PATH);
+        $cipher = $this->config->get(self::CONFIG_CIPHER_PATH);
 
         $new = $this->formatKey(
             match (Encryption::tryFrom($cipher)) {
@@ -42,10 +39,10 @@ class AppKeyGenerator implements KeyGenerator
             return $new;
         }
 
-        $this->config->set(static::CONFIG_KEY_PATH, $new);
+        $this->config->set(self::CONFIG_KEY_PATH, $new);
 
         $this->writeNewEnvironmentFileWith($write, [
-            static::ENV => [
+            self::ENV => [
                 'old' => $old,
                 'new' => $new,
             ],
