@@ -13,16 +13,14 @@ trait EnvKeySaver
     {
         $input = @file_get_contents($file);
 
-        if ($input === false) {
-            throw new RuntimeException('Error while reading environment file: ' . $file);
-        }
+        throw_if($input === false, RuntimeException::class, 'Error while reading environment file: '.$file);
 
         $match = [];
         $replacement = [];
 
         foreach ($values as $env => $value) {
             $match[] = $this->keyReplacementPattern($env, $value['old']);
-            $replacement[] = $env . '=' . $value['new'];
+            $replacement[] = $env.'='.$value['new'];
         }
 
         $replaced = preg_replace($match, $replacement, $input);
@@ -31,15 +29,13 @@ trait EnvKeySaver
             $replaced = $input;
 
             foreach ($values as $env => $value) {
-                $replaced .= "\n" . $env . '=' . $value['new'];
+                $replaced .= "\n".$env.'='.$value['new'];
             }
 
             $replaced .= "\n";
         }
 
-        if (@file_put_contents($file, $replaced) === false) {
-            throw new RuntimeException('Error while writing environment file: ' . $file);
-        }
+        throw_if(@file_put_contents($file, $replaced) === false, RuntimeException::class, 'Error while writing environment file: '.$file);
     }
 
     protected function keyReplacementPattern(string $env, string $value): string
@@ -52,6 +48,6 @@ trait EnvKeySaver
 
     protected function formatKey(string $key): string
     {
-        return 'base64:' . Base64::encode($key);
+        return 'base64:'.Base64::encode($key);
     }
 }
