@@ -14,7 +14,7 @@ function inMemoryKeyLoader(): KeyLoader
 }
 
 expect()->extend('toBeBase64', function () {
-    if (!preg_match('/^[-A-Za-z0-9+\/]+={0,3}$/', preg_quote($this->value, '/'))) {
+    if (! preg_match('/^[-A-Za-z0-9+\/]+={0,3}$/', preg_quote((string) $this->value, '/'))) {
         throw new RuntimeException(sprintf('Value %s is not a valid base64 string', $this->value));
     }
 
@@ -22,14 +22,15 @@ expect()->extend('toBeBase64', function () {
 });
 
 expect()->extend('toBeBase64NoPadding', function () {
-    if (!preg_match('/^[-A-Za-z0-9+\/]+$/', preg_quote($this->value, '/'))) {
+    if (! preg_match('/^[-A-Za-z0-9+\/]+$/', preg_quote((string) $this->value, '/'))) {
         throw new RuntimeException(sprintf('Value %s is not a valid base64 string', $this->value));
     }
+
     return $this;
 });
 
 expect()->extend('toBeBase64Url', function () {
-    if (!preg_match('/^[-A-Za-z0-9_-]+={0,3}$/', $this->value)) {
+    if (! preg_match('/^[-A-Za-z0-9_-]+={0,3}$/', (string) $this->value)) {
         throw new RuntimeException(sprintf('Value %s is not a valid base64 string', $this->value));
     }
 
@@ -37,7 +38,7 @@ expect()->extend('toBeBase64Url', function () {
 });
 
 expect()->extend('toBeBase64UrlNoPadding', function () {
-    if (!preg_match('/^[-A-Za-z0-9_-]+$/', $this->value)) {
+    if (! preg_match('/^[-A-Za-z0-9_-]+$/', (string) $this->value)) {
         throw new RuntimeException(sprintf('Value %s is not a valid base64 string', $this->value));
     }
 
