@@ -4,8 +4,8 @@ declare(strict_types=1);
 
 use CodeLieutenant\LaravelCrypto\Encoder\JsonEncoder;
 
-test('encode', function () {
-    $encoder = new JsonEncoder();
+test('encode', function (): void {
+    $encoder = new JsonEncoder;
 
     $data = ['name' => 'John Doe', 'age' => 25];
 
@@ -14,8 +14,7 @@ test('encode', function () {
     expect($encoded)->toBe('{"name":"John Doe","age":25}');
 });
 
-
-test('decode as array', function () {
+test('decode as array', function (): void {
     $encoder = new JsonEncoder(asArray: true);
 
     $data = '{"name":"John Doe","age":25}';
@@ -25,14 +24,14 @@ test('decode as array', function () {
     expect($decoded)->toBe(['name' => 'John Doe', 'age' => 25]);
 });
 
-test('decode as object', function () {
+test('decode as object', function (): void {
     $encoder = new JsonEncoder(asArray: false);
 
     $data = '{"name":"John Doe","age":25}';
 
     $decoded = $encoder->decode($data);
 
-    $class = new stdClass();
+    $class = new stdClass;
     $class->name = 'John Doe';
     $class->age = 25;
 
