@@ -4,8 +4,8 @@ declare(strict_types=1);
 
 use CodeLieutenant\LaravelCrypto\Encoder\PhpEncoder;
 
-test('encode', function () {
-    $encoder = new PhpEncoder();
+test('encode', function (): void {
+    $encoder = new PhpEncoder;
 
     $data = ['name' => 'John Doe', 'age' => 25];
 
@@ -14,9 +14,8 @@ test('encode', function () {
     expect($encoded)->toBe('a:2:{s:4:"name";s:8:"John Doe";s:3:"age";i:25;}');
 });
 
-
-test('decode', function () {
-    $encoder = new PhpEncoder();
+test('decode', function (): void {
+    $encoder = new PhpEncoder;
 
     $data = 'a:2:{s:4:"name";s:8:"John Doe";s:3:"age";i:25;}';
 
@@ -25,14 +24,14 @@ test('decode', function () {
     expect($decoded)->toBe(['name' => 'John Doe', 'age' => 25]);
 });
 
-test('decode as object', function () {
-    $encoder = new PhpEncoder();
+test('decode as object', function (): void {
+    $encoder = new PhpEncoder;
 
     $data = 'O:8:"stdClass":2:{s:4:"name";s:8:"John Doe";s:3:"age";i:25;}';
 
     $decoded = $encoder->decode($data);
 
-    $class = new stdClass();
+    $class = new stdClass;
     $class->name = 'John Doe';
     $class->age = 25;
 
