@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 use CodeLieutenant\LaravelCrypto\Support\Base64;
 
-it('encodes binary data', function () {
+it('encodes binary data', function (): void {
     $data = random_bytes(32);
 
     $encoded = Base64::encode($data);
