@@ -6,11 +6,7 @@ use Illuminate\Support\ServiceProvider;
 
 class WorkbenchServiceProvider extends ServiceProvider
 {
-    public function register(): void
-    {
-    }
+    public function register(): void {}
 
-    public function boot(): void
-    {
-    }
+    public function boot(): void {}
 }
